fix(babel): fix configuration of babel module resolver with eslint
