+++ packages/core/.eslintrc.js
@@ -1,4 +1,5 @@
 const path = require('path');
+const getModuleResolverOptions = require('./config/getModuleResolverOptions');
 
 module.exports = {
   extends: 'airbnb',
@@ -23,12 +24,13 @@ module.exports = {
   },
   settings: {
     'import/resolver': {
-      'babel-module': {},
+      'babel-module': getModuleResolverOptions('web'),
     },
   },
   globals: {},
   rules: {
     'import/no-extraneous-dependencies': [0],
+    'import/prefer-default-export': [0],
     'react/state-in-constructor': [1, 'never'],
   },
 };
