+++ packages/core/babel.config.node.js
@@ -1,15 +1,9 @@
 const path = require('path');
-const fs = require('fs');
 const { getValueFromFiles } = require('./config/irving/getValueFromFiles');
-const getServiceAliases = require('./config/irving/getServiceAliases');
+const getModuleResolverOptions = require('./config/getModuleResolverOptions');
 const {
   buildContext,
 } = require('./config/paths');
-const aliases = require('./config/aliases');
-const scopeDir = path.join(__dirname, '../');
-const packageDirs = fs.readdirSync(scopeDir);
-const packageRoots = ! packageDirs.length ? [] :
-  packageDirs.map((dir) => path.join(scopeDir, dir));
 const shimPath = path.join(
   buildContext,
   'node_modules/@irvingjs/core/utils/shimDom'
@@ -19,17 +13,7 @@ const config = {
   plugins: [
     [
       'module-resolver',
-      {
-        root: [
-          buildContext,
-          ...packageRoots,
-        ],
-        cwd: 'packagejson',
-        alias: {
-          ...getServiceAliases('node'),
-          ...aliases,
-        },
-      },
+      getModuleResolverOptions('node'),
     ],
     ['transform-globals', {
       require: {
